+++ pmxbot/slack.py
@@ -4,6 +4,8 @@ import importlib
 import logging
 import re
 import html
+import threading
+import contextlib
 
 from tempora import schedule
 
@@ -16,60 +18,36 @@ log = logging.getLogger(__name__)
 SLACK_CACHE_SECONDS = 7 * 24 * 60 * 60
 
 
-def get_ttl_hash(seconds=None):
+def iter_cursor(callable, cursor=None):
     """
-    Return the same value withing `seconds` time period
-
-    default seconds:
-        7 days == 7days*24hours*60min*60seconds == 604800 seconds
-
-    Cache time can be configured in config using `slack_cache` in seconds
+    Iterate a slack endpoint callable that uses paginated results.
     """
-    if not seconds:
-        try:
-            seconds = pmxbot.config.get('slack_cache', SLACK_CACHE_SECONDS)
-        except AttributeError:
-            # whenever pmxbot doesn't have config
-            seconds = SLACK_CACHE_SECONDS
-    return round(time.time() / seconds)
+    resp = callable(cursor=cursor)
+    yield resp.data
+    next_cursor = resp.data.get('response_metadata', {}).get('next_cursor')
+    if next_cursor:
+        yield from iter_cursor(callable, cursor=next_cursor)
 
 
 class Bot(pmxbot.core.Bot):
     def __init__(self, server, port, nickname, channels, password=None):
         token = pmxbot.config['slack token']
-        sc = importlib.import_module('slackclient')
-        self.slack = sc.SlackClient(token)
-        sr = importlib.import_module('slacker')
-        self.slacker = sr.Slacker(token)
-
+        sc = importlib.import_module('slack_sdk.rtm_v2')
+        self.slack = sc.RTMClient(token=token)
         self.scheduler = schedule.CallbackScheduler(self.handle_scheduled)
-        # Store in cache users on init
-        self.get_email_username_map(
-            ttl_hash=get_ttl_hash(pmxbot.config.get('slack_cache'))
-        )
-
-    @functools.lru_cache(maxsize=1)
-    def get_email_username_map(self, ttl_hash=get_ttl_hash()):
-        """
-        Generate a map {email -> slack username}
-
-        :param ttl_hash: hash to control the lru_cache
-        """
-        users = self.slacker.users.list()
-        if users.body.get('ok', False):
-            members = users.body.get('members', [])
-            return {
-                member.get('profile', {}).get('email', ""): member.get("name")
-                for member in members
-            }
 
     def start(self):
-        res = self.slack.rtm_connect()
-        assert res, "Error connecting"
         self.init_schedule(self.scheduler)
+        threading.Thread(target=self.run_scheduler_loop).start()
+
+        @self.slack.on("message")
+        def handle_payload(client, event):
+            self.handle_message(event)
+
+        self.slack.start()
+
+    def run_scheduler_loop(self):
         while True:
-            for msg in self.slack.rtm_read():
-                self.handle_message(msg)
             self.scheduler.run_pending()
             time.sleep(0.1)
 
@@ -86,26 +64,28 @@ class Bot(pmxbot.core.Bot):
             return
         nick = resolve_nick(msg)
 
-        channel = self.slack.server.channels.find(msg['channel']).name
-        channel = core.AugmentableMessage(channel, thread=msg.get('thread_ts'))
+        channel_name = self._get_channel_name(msg['channel'])
+        channel = core.AugmentableMessage(
+            channel_name, channel_id=msg.get('channel'), thread=msg.get('thread_ts')
+        )
 
         self.handle_action(channel, nick, html.unescape(msg['text']))
 
+    @functools.lru_cache()
+    def _get_channel_name(self, channel_id):
+        return (
+            self.slack.web_client.conversations_info(channel=channel_id)
+            .data['channel']
+            .get('name')
+        )
+
     def _resolve_nick_standard(self, msg):
-        return self.slack.server.users.find(msg['user']).name
+        return self.slack.web_client.users_info(user=msg['user']).data['user']['name']
 
     _resolve_nick_me_message = _resolve_nick_standard
 
     def _resolve_nick_bot_message(self, msg):
         return msg.get('username') or msg.get('bot_id') or 'Anonymous Bot'
-
-    def _find_user_channel(self, username):
-        """
-        Use slacker to resolve the username to an opened IM channel
-        """
-        user_id = self.slacker.users.get_user_id(username)
-        im = user_id and self.slacker.im.open(user_id).body['channel']['id']
-        return im and self.slack.server.channels.find(im)
 
     def transmit(self, channel, message):
         """
@@ -115,22 +95,55 @@ class Bot(pmxbot.core.Bot):
             If a ``thread`` attribute is present, that thread ID is used.
         :param str message: message to send.
         """
-        target = (
-            self.slack.server.channels.find(channel)
-            or self._find_user_channel(username=channel)
-            or self._find_user_channel(
-                username=self.get_email_username_map().get(channel)
-            )
-        )
         message = self._expand_references(message)
 
-        target.send_message(message, thread=getattr(channel, 'thread', None))
+        # If this action was generated from a slack message event then we should have
+        # the channel_id already. For other cases we need to query the Slack API
+        if getattr(channel, "channel_id", None):
+            channel_id = channel.channel_id
+        else:
+            channel_id = self._get_id_for_channel(channel)
+
+        self.slack.web_client.chat_postMessage(
+            channel=channel_id, text=message, thread_ts=getattr(channel, 'thread', None)
+        )
+
+    @staticmethod
+    def search_dicts(key, dicts):
+        """
+        Return the value for the first dict in dicts that has key.
+        """
+        for dict in dicts:
+            with contextlib.suppress(KeyError):
+                return dict[key]
+
+    def _get_channel_mappings(self):
+        convos = functools.partial(
+            self.slack.web_client.conversations_list,
+            exclude_archived=True,
+        )
+        return (
+            {channel['name']: channel['id'] for channel in convo['channels']}
+            for convo in iter_cursor(convos)
+        )
+
+    def _get_user_mappings(self):
+        users = functools.partial(self.slack.web_client.users_list)
+        return (
+            {user['name']: user['id'] for user in user_list['members']}
+            for user_list in iter_cursor(users)
+        )
+
+    @functools.lru_cache()
+    def _get_id_for_user(self, user_name):
+        return self.search_dicts(user_name, self._get_user_mappings())
+
+    @functools.lru_cache()
+    def _get_id_for_channel(self, channel_name):
+        return self.search_dicts(channel_name.strip('#'), self._get_channel_mappings())
 
     def _expand_references(self, message):
-        resolvers = {
-            '@': self.slacker.users.get_user_id,
-            '#': self.slacker.channels.get_channel_id,
-        }
+        resolvers = {'@': self._get_id_for_user, '#': self._get_id_for_channel}
 
         def _expand(match):
             match_type = match.groupdict()['type']